memf_correlate::traits

Trait IntoForensicEvents

pub trait IntoForensicEvents {
    // Required method
    fn into_forensic_events(self) -> Vec<ForensicEvent>;
}

Trait for types that can produce a list of ForensicEvents.

Implement this on walker output structs to enable automatic correlation and reporting.

Required Methods

fn into_forensic_events(self) -> Vec<ForensicEvent>

Consume this value and produce forensic events.